refactor(ui): rename OldTeams component file to Teams
