+++ ui/litellm-dashboard/src/app/(dashboard)/teams/page.tsx
@@ -1,9 +1,9 @@
 "use client";
 
-import OldTeams from "@/components/OldTeams";
+import Teams from "@/components/Teams";
 import useAuthorized from "@/app/(dashboard)/hooks/useAuthorized";
 
 export default function TeamsPage() {
   const { accessToken, userId, userRole, premiumUser } = useAuthorized();
-  return <OldTeams accessToken={accessToken} userID={userId} userRole={userRole} premiumUser={premiumUser ?? false} />;
+  return <Teams accessToken={accessToken} userID={userId} userRole={userRole} premiumUser={premiumUser ?? false} />;
 }
